+++ packages/moonstone/Panels/ApplicationCloseButton.js
@@ -23,6 +23,15 @@ const ApplicationCloseButton = kind({
 
 	propTypes: /** @lends moonstone/Panels.ApplicationCloseButton.prototype */ {
 		/**
+		 * The background-color opacity of this button; valid values are `'opaque'`, `'translucent'`,
+		 * and `'transparent'`.
+		 *
+		 * @type {String}
+		 * @default 'transparent'
+		 * @public
+		 */
+		backgroundOpacity: PropTypes.oneOf(['opaque', 'translucent', 'transparent']),
+		/**
 		 * A function to run when app close button is clicked
 		 *
 		 * @type {Function}
@@ -30,19 +39,23 @@ const ApplicationCloseButton = kind({
 		onApplicationClose: PropTypes.func
 	},
 
+	defaultProps: {
+		backgroundOpacity: 'transparent'
+	},
+
 	styles: {
 		css,
 		className: 'applicationCloseButton'
 	},
 
-	render: ({onApplicationClose, ...rest}) => {
+	render: ({backgroundOpacity, onApplicationClose, ...rest}) => {
 		return (
 			<IconButton
 				{...rest}
 				aria-label={$L('Exit app')}
-				small
-				backgroundOpacity="transparent"
+				backgroundOpacity={backgroundOpacity}
 				onClick={onApplicationClose}
+				small
 			>
 				closex
 			</IconButton>